__init__.py cleanup
- moved routes to new `routes.py` file
- moved helper functions to new `pb_mock.py` file
